Fix Card ID feature
Now instead of just copying the card ID (e.g. '#16') we'll copy the link
to that card. This is due to Trello not supporting automatic replacement
of these IDs.

The feature basically stays the same, but now it works again. Paste the
link into a comment or description and Trello will replace it with a
reference!

Close #11